An attack
I put an attack into the game.For the sake of this the sprites are not
rendered for now because there's no animation so all you see is the
hurtboxes, collisionboxes and hitboxes. Move around with A and D and
attack with F.You will see the hitbox appear and disappear.