Move hit-/hurtboxes to states and turning work
Moved all kinds of boxes to states since their layout will vary greatly
based on whether the character is ducking,jumping,standing or performing
an attack.

Made sure that input handling and such works when the character is on
the right side of the screen looking left

Player 1 is now controlled by keyboard, player 2 by Gamepad.